Fix ExchangeRateReader.java
We added a few imports so that Java can recognize some of the types that
we are using in our code. In addition, we changed some of the types that
we were using in order to actually retrieve the values that we want the
tests to check. For example, we changed getExchangeRate to obtain a
JSONObject instead of a JSONArray.

Disclaimer: There are some errors that are resulting in one of the tests
not passing. For instance, an incorrect exchange rate value is being
retrieved. This will need to be fixed before Wednesday night.

Co-authored-by: @BananaObserved <golds349@morris.umn.edu>